fix: alert shows duplicate attribute of Repeater
Update and small refactor with split big utils file
Repair of data attributes, now is not part of for list in alert dialog

JIRA: F1-870
