+++ libs/sdk-ui-dashboard/src/presentation/widget/insight/configuration/InsightAlertConfig/hooks/useInsightAlerting.ts
@@ -5,6 +5,7 @@ import {
     IInsightWidget,
 } from "@gooddata/sdk-model";
 import { useToastMessage } from "@gooddata/sdk-ui-kit";
+import { fillMissingTitles, useBackendStrict, useWorkspaceStrict } from "@gooddata/sdk-ui";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { v4 as uuidv4 } from "uuid";
 import {
@@ -30,20 +31,21 @@ import {
     selectEntitlementMaxAutomationRecipients,
     selectExecutionResultByRef,
     selectEnableAlertAttributes,
+    selectCatalogAttributes,
+    selectCatalogDateDatasets,
     selectSeparators,
 } from "../../../../../../model/index.js";
-import {
-    createDefaultAlert,
-    getSupportedInsightAttributesByInsight,
-    getSupportedInsightMeasuresByInsight,
-} from "../utils.js";
-import { messages } from "../messages.js";
-import { useSaveAlertToBackend } from "./useSaveAlertToBackend.js";
-import { fillMissingTitles, useBackendStrict, useWorkspaceStrict } from "@gooddata/sdk-ui";
 import { convertCurrentUserToAutomationRecipient } from "../../../../../../_staging/automation/index.js";
 import { useMetricsAndFacts } from "../../../../../../_staging/sharedHooks/useMetricsAndFacts.js";
 import { DEFAULT_MAX_RECIPIENTS } from "../../../../../scheduledEmail/DefaultScheduledEmailDialog/constants.js";
-import { useAttributes } from "../../../../../../_staging/sharedHooks/useAttributes.js";
+import { messages } from "../messages.js";
+import {
+    getSupportedInsightAttributesByInsight,
+    getSupportedInsightMeasuresByInsight,
+} from "../utils/items.js";
+import { createDefaultAlert } from "../utils/convertors.js";
+
+import { useSaveAlertToBackend } from "./useSaveAlertToBackend.js";
 
 type InsightWidgetAlertingViewMode = "list" | "edit" | "create";
 
@@ -79,6 +81,8 @@ export const useInsightWidgetAlerting = ({ widget, closeInsightWidgetMenu }: IIn
     );
     const canManageAttributes = useDashboardSelector(selectEnableAlertAttributes);
     const canCreateAutomation = useDashboardSelector(selectCanCreateAutomation);
+    const catalogAttributes = useDashboardSelector(selectCatalogAttributes);
+    const catalogDateDatasets = useDashboardSelector(selectCatalogDateDatasets);
     const currentUser = useDashboardSelector(selectCurrentUser);
     const destinations = useDashboardSelector(selectNotificationChannels);
     const canManageAutomations = useDashboardSelector(selectCanManageWorkspace);
@@ -137,7 +141,14 @@ export const useInsightWidgetAlerting = ({ widget, closeInsightWidgetMenu }: IIn
         insight,
     });
     const { metricsAndFacts, metricsAndFactsLoading, metricsAndFactsLoadingError } = useMetricsAndFacts();
+    const { metrics: catalogMetrics } = metricsAndFacts ?? { metrics: [] };
     const locale = useDashboardSelector(selectLocale);
+
+    const supportedAttributes = useMemo(
+        () => getSupportedInsightAttributesByInsight(insight, catalogDateDatasets),
+        [insight, catalogDateDatasets],
+    );
+
     const supportedMeasures = useMemo(
         () =>
             getSupportedInsightMeasuresByInsight(
@@ -146,13 +157,6 @@ export const useInsightWidgetAlerting = ({ widget, closeInsightWidgetMenu }: IIn
         [insight, locale],
     );
     const defaultMeasure = supportedMeasures[0];
-
-    const supportedAttributes = useMemo(() => getSupportedInsightAttributesByInsight(insight), [insight]);
-    const displayForms = useMemo(
-        () => supportedAttributes.map((attribute) => attribute.attribute.attribute.displayForm),
-        [supportedAttributes],
-    );
-    const { attributes, attributesLoading, attributesLoadingError } = useAttributes(displayForms);
 
     const defaultNotificationChannelId = destinations[0]?.id;
     const hasAlerts = alerts.length > 0;
@@ -176,7 +180,6 @@ export const useInsightWidgetAlerting = ({ widget, closeInsightWidgetMenu }: IIn
             defaultMeasure &&
             defaultNotificationChannelId &&
             !metricsAndFactsLoading &&
-            !attributesLoading &&
             !defaultAlert
         ) {
             setDefaultAlert(
@@ -189,10 +192,7 @@ export const useInsightWidgetAlerting = ({ widget, closeInsightWidgetMenu }: IIn
                     metricsAndFacts?.metrics ?? [],
                 ),
             );
-        } else if (
-            (widgetFiltersStatus === "error" || metricsAndFactsLoadingError || attributesLoadingError) &&
-            !defaultAlert
-        ) {
+        } else if ((widgetFiltersStatus === "error" || metricsAndFactsLoadingError) && !defaultAlert) {
             closeInsightWidgetMenu();
             addError(messages.alertLoadingError);
         }
@@ -207,10 +207,9 @@ export const useInsightWidgetAlerting = ({ widget, closeInsightWidgetMenu }: IIn
         metricsAndFacts,
         metricsAndFactsLoading,
         metricsAndFactsLoadingError,
-        attributesLoading,
-        attributesLoadingError,
         addError,
         currentUser,
+        users,
     ]);
 
     useEffect(() => {
@@ -328,8 +327,7 @@ export const useInsightWidgetAlerting = ({ widget, closeInsightWidgetMenu }: IIn
             isLoadingFilters ||
             isRefreshingAutomations ||
             isDeletingAlert ||
-            metricsAndFactsLoading ||
-            attributesLoading,
+            metricsAndFactsLoading,
         destinations,
         users,
         alerts,
@@ -354,7 +352,8 @@ export const useInsightWidgetAlerting = ({ widget, closeInsightWidgetMenu }: IIn
         maxAutomationsRecipients,
         canManageAttributes,
         canCreateAutomation,
-        catalogMeasures: metricsAndFacts?.metrics ?? [],
-        catalogAttributes: attributes ?? [],
+        catalogMeasures: catalogMetrics ?? [],
+        catalogAttributes: catalogAttributes ?? [],
+        catalogDateDatasets: catalogDateDatasets ?? [],
     };
 };
