+++ libs/sdk-ui-dashboard/src/presentation/widget/insight/configuration/InsightAlertConfig/utils/guards.ts
@@ -1,0 +1,48 @@
+// (C) 2022-2024 GoodData Corporation
+import { IAutomationAlert, IAutomationMetadataObject } from "@gooddata/sdk-model";
+
+import { ARITHMETIC_OPERATORS } from "../constants.js";
+
+/**
+ * @internal
+ */
+export function isChangeOperator(alert?: IAutomationAlert): boolean {
+    if (alert?.condition?.type === "relative") {
+        return alert.condition.measure.operator === ARITHMETIC_OPERATORS.ARITHMETIC_OPERATOR_CHANGE;
+    }
+    return false;
+}
+
+/**
+ * @internal
+ */
+export function isDifferenceOperator(alert?: IAutomationAlert): boolean {
+    if (alert?.condition?.type === "relative") {
+        return alert.condition.measure.operator === ARITHMETIC_OPERATORS.ARITHMETIC_OPERATOR_DIFFERENCE;
+    }
+    return false;
+}
+
+/**
+ * @internal
+ */
+export function isChangeOrDifferenceOperator(alert?: IAutomationAlert): boolean {
+    return isChangeOperator(alert) || isDifferenceOperator(alert);
+}
+
+/**
+ * @internal
+ */
+export function isAlertValueDefined(alert?: IAutomationAlert): boolean {
+    if (alert?.condition?.type === "relative") {
+        return typeof alert.condition.threshold !== "undefined";
+    }
+    return typeof alert?.condition?.right !== "undefined";
+}
+
+/**
+ * @internal
+ */
+export function isAlertRecipientsValid(alert?: IAutomationMetadataObject): boolean {
+    return !!alert?.recipients?.length;
+}
